+++ CHANGES.md
@@ -1,5 +1,35 @@
 # Changes
 
+## edge-23.12.4
+
+This edge release includes fixes and improvements to the destination
+controller's endpoint resolution API.
+
+* Fixed an issue in the control plane where discovery for pod IP addresses could
+  hang indefinitely ([#11815])
+* Updated the proxy to enforce time limits on control plane response streams so
+  that proxies more naturally distribute load over control plane replicas
+  ([#11837])
+* Fixed the policy's controller service metadata responses so that proxy logs
+  and metrics have informative values ([#11842])
+
+[#11842]: https://github.com/linkerd/linkerd2/pull/11842
+[#11837]: https://github.com/linkerd/linkerd2/pull/11837
+[#11815]: https://github.com/linkerd/linkerd2/pull/11815
+
+## edge-23.12.3
+
+This edge release contains improvements to the logging and diagnostics of the
+destination controller.
+
+* Added a control plane metric to count errors talking to the Kubernetes API
+  ([#11774])
+* Fixed an issue causing spurious destination controller error messages for
+  profile lookups on unmeshed pods with port in default opaque list ([#11550])
+
+[#11774]: https://github.com/linkerd/linkerd2/pull/11774
+[#11550]: https://github.com/linkerd/linkerd2/pull/11550
+
 ## edge-23.12.2
 
 This edge release includes a restructuring of the proxy's balancer along with